Added check on differences of to/from bounded_normal_rh probit in quantile_distributions_mod. While not guaranteeing correctness, this can catch all but unusual cancelling errors in the transforms when turned on.
Added namelist control for three options in quantile_distributions_mod.

Changed the algorithm for determining when to switch to uniform tails for
the bounded_normal_rh.

Switched the normal_distributions module to always do a high accuracy
inverse cdf for now. This supports the tests in the quantile_distribution_mo