creates a regular grid with valid data and writes it to a netcdf file
note that since we're using lon/lat for the indices, ncview tries to
put continent boundaries on the plot which is annoying at low resolution.
use 'opts' and 'no boundaries' and 'ok' to turn that off.

it also makes a grid based on a caminput.nc file and tries to interpolate
to that grid.  it's putting some data on it but clearly in the wrong place.
still working on the interp calls but this is progress.