make resolution a namelist variable.  it controls the density of the origin grid.
currently testing with 1 degree grids.  for ncview, set bi-lin to repl, and
if there are large blobby white squares when looking at field0, use opts to
turn coastlines off.  the diff for the case that's currently selected in the
input.nml has anomalies at lon 0 and 180, and at lat 90 that i will look into
more as soon as i get a chance.