make clear obs_window_seconds and obs_window_days are currently unsupported
both filter and perfect_model_obs have obs_window_days and obs_window_seconds
in their namelists.  these are for future use and currently unsupported.  update
the docs to be more clear about this and update the code to print out a warning
message if they are changed to be >= 0.

also check both the days and seconds values for obs_window_xxx and also
init_time_xxx.  if either are >= 0, call set_time().
that will catch and print a fatal error if one is set and the other is left at -1.
with the old code, if someone set seconds >= 0 but left days negative, the init
time would have been set to 0, 0 and the namelist values wouldn't be used.