fix the longitudes and handle poles differently based on test results
update the quad tests, and fix known problems in the quad interpolate code.
there are still some quads along the lon=0 boundary high in latitudes that
seem to have bad interpolation values but they can be corrected to not be
outside of the original corners of the quads.  the tests on the rest of the
grids look fine.  the pole issue seems best handled by setting "pole_wrap" to
be false for the U grid and then the pole test is used and pole points are
rejected.