Convert original Aether variable names to new
This matlab script creates new block restart files
with the names Aaron provided in mid-January, 2024.
The data are the same as the input (old names) files.
This is a commit of the the development script,
which converts only 2 neutrals and 2 ions for the
requested number of block(s) and member(0-based).