Enhanced Performance
Changes to the model_mod and noah_hydro_mod to enhance performance
when running a full CONUS domain. Various parts of the noah_hydro_mod
code are rearranged to avoid extra computations that are not needed
when the LSM is turned off. Also, the link tree structure is modified
such that only the individual upstream links from any reach are stored.
We also limit the number of upstream links to '5'. This helps alleviate
any issues in the Routelink file (which can make the stream network unphysical).
This change makes our Along-The-Stream (ATS) Localization runs a lot faster.

Changes in the model_mod makes the code faster when distributing the close
by reaches to each individual task. Also, there is a change that allows reading
climatology files when a hybrid EnKF variant is invoked.