Enabled running *_to_* outside of the restart directory
Added aether_restart_dirname to all filenames that are opened.
Removed aether_restart_dirname (and local names) from argument lists
because it's a global variable.
Updated readme.rst with better instructions for testing *_to_*
and clarifications.
Made the default aether_restart_dirname = '.' so that it can be
left out of the namelist when running the executables
in a run directory where the restart files are.

Passed file conversion tests and model_mod_check tests 1-5,7.