Station location is now a namelist option
Removed the hard-coded location of BATS station
and added it as `station_location` in the input.nml
This allows the users to run the code for any location
on earth (i.e., it doesn't have to be Bermuda station).

I also cleaned up the input.nml and stripped out the
old path names to Robin's directories.