Modifications to the obs_converter
Updated the BATS station location in the obs_converter.
Previously it assumed a lon of 64 while it should be
64W or 296.
Also, the BATS text data file has been modified and
that required changing the input parameters for
reading the quantities properly.

filter has been run several times and obs_seq.final
was assessed with no apparent issues.