query the fill value from the mom6.static.nc file
mask where geolon is the fillvalue.  Note the mask for geolon/lat/u/v
is not the same as 'wet' from the ocean_geometry.nc file.

Setting the masked lat lon values to a land value. This is because
set_location is used on all the lat, lon vaues and 1e20 will blow up
set_location