Existing code did not take appropriate action when the bnrh transform failed because the sample ensemble covariance was not positive. The existing code simply returned the input ensemble but did not indicate that an error had occurred. This resulted in the untransformed ensemble being used as if it had been transformed in filter_mod.f90 for inflation and in assim_tools_mod.f90 for assimilation. In general, this was not a problem because zero covariance in the ensemble led to both inflation and assimilation doing nothing. However, due to numerical round-off, it was possible to get cases where the inflation could result in violating the bounds of the bnrh distribution. This commit attempts to correct this behavior by updates to probit_transform_mod.f90, filter_mod.f90, and assim_tools_mod.f90.
probit_transform_mod.f90:
An error return argument was added to subroutine to_probit_bounded_normal_rh. It returns a 1 if the
ensemble standard deviation was not positive (failure) and a 0 otherwise. This returned argument is
passed up through transform_to_probit and transform_all_to_probit, where it becomes an array with one
entry for each state variable being transformed. In the case of an error, the returned ‘transformed’
ensemble is simply the input original ensemble. The check for a non-positive standard deviation was
eliminated in the case where ‘use_input_p’ is true since this should never be called if the original
transformation failed. Subroutine transform_all_from_probit was modified to include an additional
logical array that indicates whether each of the state variables to be transformed was successfully
transformed to probit space. The inverse transform is only computed for the state variables for which
the original transform was successful.

filter_mod.f90:
All changes are in subroutine filter_ensemble_inflate. The inflation is only performed if the error
return from the transform_to_probit is 0. Note that this is called for each group if a group filter is
being run and it would be possible for some groups to transform successfully while others failed. This
case is handled properly.

assim_tools_mod.f90:
Two logical arrays were added to keep track of which transforms fail for the observed variables
(obs_probit_trans_ok) and state variables (state_probit_trans_ok); these are allocated on entry into
subroutine filter_assim and deallocated on exit. The result of the transforms is only copied into the
appropriate ensemble_handle storage if the transform was successful. When observations are sequentially
assimilated, the observation ensemble is only converted back to physical space if its original transform
was successful. The observation prior and posterior are transformed back to probit space in the
SEQUENTIAL_OBS loop. If this transform fails for any group, the loop is cycled for this observation so
that it has no impact. In the STATE_UPDATE loop, if the transform to probit space for the current state
variable failed, the loop is cycled. The same is done in the OBS_UPDATE loop. Note that groups are not
relevant to this step. Finally, when transform_all_from_probit is called for state variables, only those
that were successfully transformed originally are transformed back.