array version of quad_idw_interp from branch idw_cice-jla
checks and errors out if value is out-of-bounds
clamps also after error check
same subroutine signature (also same as quad_utils_mod), no ensemble

Original commits messages from idw_cice-jla:
Changed quad_idw_interpolate to a function. Added definition of minimum distance to
avoid round-off. Added error if interpolant is outside of range of grid point
values. The error return is a short-term measure to see if this is really happening.
Once we understand this better can probably change to just constraining result
to be within the range.

Converted quad_idw_interp to do the whole ensemble since all of the distance computations
are identical for each member. Added the code to correct values that are outside of the
range of the gridpoint values. This code would never be called in this version because
the error check dies before getting to it.
@jlaucar

Co-authored-by: jlaucar <jla@ucar.edu>