Added in tolerance exponent as a namelist settable feature. Tuned this value from 0.75 to 0.52 to avoid previous failure to converge error messages.
This version passes the normal, beta and gamma developers tests for the
following compilers on Derecho with the standard mkmf files:
gcc 12.4.0
gcc 12.2.0
gcc 13.2.0
ifort 2024.2
ifx 2025.0.3
nvidia 24.11

It passed all tests except the second set of convergence tests for the
beta distribution for the Cray cce 17.0.1. That compiler crashed for the
second part of the beta tests, likely due to a compiler error.

It passed all three tests on a mac with gfortran 13.3.0 with -O0, -O1,
-O2 and -O3.