change parse_args_test to be more automated.
renamed the original test to interactive_parse_args_test.  it read standard input
and prints out the values it gets.  there is still a parse_args_test.in text file
that can be used as input, or typing from the termina.

the new parse_args_test uses the fortran-testanything module which prints 'ok'
for each test that passes.  this version of the test needs no input files.